move v1 tests to separate folder
