+++ src/tests/v1/contact.routes.spec.js
@@ -2,8 +2,8 @@ import chai from 'chai'
 import request from 'supertest'
 import randomWords from 'random-words'
 import uuid from 'uuid4'
-import { Login } from '../utils/login'
-import app from '..'
+import { Login } from '../../utils/login'
+import app from '../..'
 
 const VERSION = '1';
 const { expect } = chai
