+++ src/tests/v1/userRole.routes.spec.js
@@ -1,10 +1,13 @@
-/* eslint-disable no-unused-vars */
+import chai from 'chai'
 import request from 'supertest'
-import { Login } from '../utils/login'
-import app from '..'
+import { Login } from '../../utils/login'
+import app from '../..'
 
 const VERSION = '1';
-describe('CSV Dump Negative Tests', function() {
+const { expect, assert } = chai
+const role = { role: 'test' + Date.now(), path: '/test', method: 'GET' }
+
+describe('User roles positive tests', function() {
   const authed = new Login(VERSION)
   let token
 
@@ -15,81 +18,73 @@ describe('CSV Dump Negative Tests', function() {
   after(async function() {
     await authed.destroyToken()
   })
-  it('Negative Test (400) for CSV Dump on Contact', function(done) {
+
+  it('should create a user role', function(done) {
     request(app)
-      .get(`/v${VERSION}/csv/contact`)
+      .post(`/v${VERSION}/userRole`)
+      .send(role)
+      .set('Accept', 'application/json')
+      .set('token', token)
+      .expect('Content-Type', 'text/html; charset=utf-8')
+      .expect(200)
+      .end((err, res) => {
+        if (err) return done(err)
+        expect(res.text).to.equal('policy created')
+        done()
+      })
+  })
+  it('should get all user roles', function(done) {
+    request(app)
+      .get(`/v${VERSION}/userRole`)
+      .set('Accept', 'application/json')
+      .set('token', token)
+      .expect('Content-Type', 'application/json; charset=utf-8')
+      .expect(200)
+      .end((err, res) => {
+        if (err) return done(err)
+        assert.isObject(res.body, 'body is an object')
+        done()
+      })
+  })
+  it('should delete a user role', function(done) {
+    request(app)
+      .post(`/v${VERSION}/userRole/delete`)
+      .send(role)
+      .set('Accept', 'application/json')
+      .set('token', token)
+      .expect('Content-Type', 'text/html; charset=utf-8')
+      .expect(200)
+      .end((err, res) => {
+        if (err) return done(err)
+        expect(res.text).to.equal('Policy deleted')
+        done()
+      })
+  })
+
+  it('should not create a user role', function(done) {
+    request(app)
+      .post(`/v${VERSION}/userRole`)
       .set('Accept', 'application/json')
       .set('token', token)
       .expect('Content-Type', 'text/html; charset=utf-8')
       .expect(400)
       .end((err, res) => {
         if (err) return done(err)
+        expect(res.text).to.equal('Bad Request')
         done()
       })
   })
-
-  it('Negative Test (400) for CSV Dump on Entity', function(done) {
+  it('should not delete a user', function(done) {
     request(app)
-      .get(`/v${VERSION}/csv/entity`)
+      .post(`/v${VERSION}/userRole/delete`)
       .set('Accept', 'application/json')
       .set('token', token)
       .expect('Content-Type', 'text/html; charset=utf-8')
       .expect(400)
       .end((err, res) => {
         if (err) return done(err)
+        expect(res.text).to.equal('Request does not inclue required parameters')
         done()
       })
   })
-
-  it('Negative Test (400) for CSV Dump on User', function(done) {
-    request(app)
-      .get(`/v${VERSION}/csv/user`)
-      .set('Accept', 'application/json')
-      .set('token', token)
-      .expect('Content-Type', 'text/html; charset=utf-8')
-      .expect(400)
-      .end((err, res) => {
-        if (err) return done(err)
-        done()
-      })
-  })
-
-  it('Negative Test (400) for CSV Dump on UserRole', function(done) {
-    request(app)
-      .get(`/v${VERSION}/csv/userRole`)
-      .set('Accept', 'application/json')
-      .set('token', token)
-      .expect('Content-Type', 'text/html; charset=utf-8')
-      .expect(400)
-      .end((err, res) => {
-        if (err) return done(err)
-        done()
-      })
-  })
-
-  it('Negative Test (404) for CSV Dump endpoint', function(done) {
-    request(app)
-      .get(`/v${VERSION}/csv/`)
-      .set('Accept', 'application/json')
-      .set('token', token)
-      .expect('Content-Type', 'text/html; charset=utf-8')
-      .expect(404)
-      .end((err, res) => {
-        if (err) return done(err)
-        done()
-      })
-  })
-
-  // it('Negative Test (503) for CSV Dump endpoint', (done) => {
-  //   request(app)
-  //     .get('/csv/%^&&!@&#)(@*&#()@*&)(*&)(*&(@&#)(!&*#)(*&!@#()*&()*!@&#()*&#&))*&(&()*&#!!@#!@#!@#)(&*)(*&)(*!&#')
-  //     .set('Accept', 'application/json')
-  //     .set('token', token)
-  //     .expect('Content-Type', 'text/html; charset=utf-8')
-  //     .expect(503)
-  //     .end((err, res) => {
-  //       if (err) return done(err)
-  //       done()
-  //     })
-  // })
 })
